ml-api/requirements.txt
fastapi
uvicorn
torch
torchvision
pillow
numpy
pydicom
requests
tqdm
matplotlib
scikit-learn
httpx
